requirements.txt
python-dotenv>=1.0.0
requests>=2.28.0
openai>=1.0.0
azure-ai-formrecognizer>=3.2.1
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
streamlit>=1.33.0
gradio>=4.20.0
beautifulsoup4>=4.12.0
tiktoken>=0.6.0
pydantic>=1.10.0
